backend/requirements.txt
asgiref==3.8.1
cffi==1.17.1
cryptography==44.0.2
Django==5.2
djangorestframework==3.16.0
jwt==1.3.1
pycparser==2.22
sqlparse==0.5.3
